refactor #86: Separates components and logic for improved reusability and readability
