+++ frontend/src/utils/formatters.ts
@@ -148,3 +148,44 @@ export const formatDateKorean = (date: string | Date): string => {
 export const formatPrice = (price: number): string => {
   return new Intl.NumberFormat("ko-KR").format(price) + "원";
 };
+
+/**
+ * 사업자등록번호 포맷팅 (예: "123-45-67890")
+ */
+export const formatBusinessNumber = (value: string): string => {
+  const numbers = value.replace(/-/g, "").slice(0, 10);
+  if (numbers.length <= 3) return numbers;
+  if (numbers.length <= 5)
+    return `${numbers.slice(0, 3)}-${numbers.slice(3)}`;
+  return `${numbers.slice(0, 3)}-${numbers.slice(3, 5)}-${numbers.slice(5)}`;
+};
+
+/**
+ * 전화번호 포맷팅 (예: "010-1234-5678")
+ */
+export const formatPhoneNumber = (value: string): string => {
+  const numbers = value.replace(/\D/g, "");
+  
+  if (numbers.length === 0) return "";
+  if (numbers.length <= 2) return numbers;
+  
+  if (numbers.startsWith("02")) {
+    if (numbers.length <= 5) return `${numbers.slice(0, 2)}-${numbers.slice(2)}`;
+    if (numbers.length <= 9) return `${numbers.slice(0, 2)}-${numbers.slice(2, 5)}-${numbers.slice(5)}`;
+    return `${numbers.slice(0, 2)}-${numbers.slice(2, 6)}-${numbers.slice(6, 10)}`;
+  }
+  
+  if (numbers.length === 8) {
+    return `${numbers.slice(0, 4)}-${numbers.slice(4)}`;
+  }
+  
+  if (numbers.length >= 3) {
+    const prefix = numbers.slice(0, 3);
+    if (numbers.length <= 6) return `${prefix}-${numbers.slice(3)}`;
+    if (numbers.length <= 10) return `${prefix}-${numbers.slice(3, 6)}-${numbers.slice(6)}`;
+    return `${prefix}-${numbers.slice(3, 7)}-${numbers.slice(7, 11)}`;
+  }
+  
+  if (numbers.length <= 7) return `${numbers.slice(0, 3)}-${numbers.slice(3)}`;
+  return `${numbers.slice(0, 3)}-${numbers.slice(3, 7)}-${numbers.slice(7, 11)}`;
+};
